requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
PyMySQL==1.1.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.6.0
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
platformdirs==3.3.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.22.0
virtualenvwrapper-win==1.2.7
xlrd==2.0.1
asgiref==3.6.0
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
Pillow==10.0.0
platformdirs==3.3.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.22.0
virtualenvwrapper-win==1.2.7
xlrd==2.0.1